Split the `sFnRegistryConfiguration` function (#58)
* split registry sFn into two smaller state functions

* improve storage tests

* add registry tests
