+++ components/operator/internal/state/initialize.go
@@ -14,5 +14,5 @@ func sFnInitialize(_ context.Context, _ *reconciler, s *systemState) (stateFn, *
 		return nextState(sFnDeleteResources)
 	}
 
-	return nextState(sFnRegistryConfiguration)
+	return nextState(sFnAccessConfiguration)
 }
