+++ components/operator/internal/state/registry_test.go
@@ -6,15 +6,17 @@ import (
 
 	"github.com/kyma-project/docker-registry/components/operator/api/v1alpha1"
 	"github.com/kyma-project/docker-registry/components/operator/internal/chart"
+	"github.com/kyma-project/docker-registry/components/operator/internal/registry"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
+	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 )
 
-func Test_sFnRegistryConfiguration(t *testing.T) {
-	t.Run("internal registry using default storage", func(t *testing.T) {
+func Test_sFnAccessConfiguration(t *testing.T) {
+	t.Run("setup node port only when registry secret does not exist", func(t *testing.T) {
 		s := &systemState{
 			instance:       v1alpha1.DockerRegistry{},
 			statusSnapshot: v1alpha1.DockerRegistryStatus{},
@@ -25,140 +27,84 @@ func Test_sFnRegistryConfiguration(t *testing.T) {
 			log: zap.NewNop().Sugar(),
 		}
 		expectedFlags := map[string]interface{}{
-			"configData": map[string]interface{}{
-				"storage": map[string]interface{}{
-					"filesystem": map[string]interface{}{
-						"rootdirectory": "/var/lib/registry",
-					},
-				},
-			},
-			"storage":          "filesystem",
 			"registryNodePort": int64(32_137),
 		}
 
-		next, result, err := sFnRegistryConfiguration(context.Background(), r, s)
+		next, result, err := sFnAccessConfiguration(context.Background(), r, s)
 		require.NoError(t, err)
 		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
+		requireEqualFunc(t, sFnStorageConfiguration, next)
 
 		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
 		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
 	})
 
-	t.Run("internal registry using azure storage", func(t *testing.T) {
+	t.Run("setup node port and use existing username and password", func(t *testing.T) {
+		registrySecret := &corev1.Secret{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      registry.SecretName,
+				Namespace: "kyma",
+				Labels: map[string]string{
+					registry.LabelConfigKey: registry.LabelConfigVal,
+				},
+			},
+			Data: map[string][]byte{
+				"isInternal": []byte("true"),
+				"username":   []byte("ala"),
+				"password":   []byte("makota"),
+			},
+		}
+
+		registryDeploy := &appsv1.Deployment{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      registry.DeploymentName,
+				Namespace: "kyma",
+			},
+			Spec: appsv1.DeploymentSpec{
+				Template: corev1.PodTemplateSpec{
+					Spec: corev1.PodSpec{
+						Containers: []corev1.Container{
+							{
+								Env: []corev1.EnvVar{
+									{
+										Name:  registry.HttpEnvKey,
+										Value: "httpEnvKeyVal",
+									},
+								},
+							},
+						},
+					},
+				},
+			},
+		}
+
 		s := &systemState{
 			instance: v1alpha1.DockerRegistry{
 				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kyma-system",
-				},
-				Spec: v1alpha1.DockerRegistrySpec{
-					Storage: &v1alpha1.Storage{
-						Azure: &v1alpha1.StorageAzure{
-							SecretName: "azureSecret",
-						},
-					},
+					Namespace: "kyma",
 				},
 			},
 			statusSnapshot: v1alpha1.DockerRegistryStatus{},
 			flagsBuilder:   chart.NewFlagsBuilder(),
 		}
 		r := &reconciler{
-			k8s: k8s{client: fake.NewClientBuilder().Build()},
+			k8s: k8s{client: fake.NewClientBuilder().WithObjects(registrySecret, registryDeploy).Build()},
 			log: zap.NewNop().Sugar(),
 		}
-		azureSecret := &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "azureSecret",
-				Namespace: "kyma-system",
-			},
-			Data: map[string][]byte{
-				"accountName": []byte("accountName"),
-				"accountKey":  []byte("accountKey"),
-				"container":   []byte("container"),
-			},
-		}
-		require.NoError(t, r.k8s.client.Create(context.Background(), azureSecret))
-
 		expectedFlags := map[string]interface{}{
-			"storage": "azure",
-			"secrets": map[string]interface{}{
-				"azure": map[string]interface{}{
-					"accountName": "accountName",
-					"accountKey":  "accountKey",
-					"container":   "container",
-				},
-			},
 			"registryNodePort": int64(32_137),
+			"dockerRegistry": map[string]interface{}{
+				"username": "ala",
+				"password": "makota",
+			},
+			"registryHTTPSecret": "httpEnvKeyVal",
+			"rollme":             "dontrollplease",
 		}
 
-		next, result, err := sFnRegistryConfiguration(context.Background(), r, s)
+		next, result, err := sFnAccessConfiguration(context.Background(), r, s)
 		require.NoError(t, err)
 		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
-
-		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
-		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
-	})
-	t.Run("internal registry using s3 storage", func(t *testing.T) {
-		s := &systemState{
-			instance: v1alpha1.DockerRegistry{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "kyma-system",
-				},
-				Spec: v1alpha1.DockerRegistrySpec{
-					Storage: &v1alpha1.Storage{
-						S3: &v1alpha1.StorageS3{
-							Bucket:         "bucket",
-							Region:         "region",
-							RegionEndpoint: "regionEndpoint",
-							Encrypt:        false,
-							Secure:         true,
-							SecretName:     "s3Secret",
-						},
-					},
-				},
-			},
-			statusSnapshot: v1alpha1.DockerRegistryStatus{},
-			flagsBuilder:   chart.NewFlagsBuilder(),
-		}
-		r := &reconciler{
-			k8s: k8s{client: fake.NewClientBuilder().Build()},
-			log: zap.NewNop().Sugar(),
-		}
-		azureSecret := &corev1.Secret{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "s3Secret",
-				Namespace: "kyma-system",
-			},
-			Data: map[string][]byte{
-				"accessKey": []byte("accessKey"),
-				"secretKey": []byte("secretKey"),
-			},
-		}
-		require.NoError(t, r.k8s.client.Create(context.Background(), azureSecret))
-
-		expectedFlags := map[string]interface{}{
-			"storage": "s3",
-			"s3": map[string]interface{}{
-				"bucket":         "bucket",
-				"region":         "region",
-				"regionEndpoint": "regionEndpoint",
-				"encrypt":        false,
-				"secure":         true,
-			},
-			"secrets": map[string]interface{}{
-				"s3": map[string]interface{}{
-					"accessKey": "accessKey",
-					"secretKey": "secretKey",
-				},
-			},
-			"registryNodePort": int64(32_137),
-		}
-
-		next, result, err := sFnRegistryConfiguration(context.Background(), r, s)
-		require.NoError(t, err)
-		require.Nil(t, result)
-		requireEqualFunc(t, sFnConfigurationStatus, next)
+		requireEqualFunc(t, sFnStorageConfiguration, next)
 
 		require.EqualValues(t, expectedFlags, s.flagsBuilder.Build())
 		require.Equal(t, v1alpha1.StateProcessing, s.instance.Status.State)
